feat(RHOAIENG-65107): move test-facing enums to plugin-core/model-serving-labels
