+++ packages/cypress/cypress/tests/e2e/dataScienceProjects/models/testModelTokenAuth.cy.ts
@@ -1,7 +1,7 @@
 import {
   ModelLocationSelectOption,
   ModelTypeLabel,
-} from '@odh-dashboard/model-serving/types/form-data';
+} from '@odh-dashboard/plugin-core/model-serving-labels';
 import { projectListPage, projectDetails } from '../../../../pages/projects';
 import {
   modelServingGlobal,
