+++ packages/cypress/cypress/tests/e2e/modelsAsAService/testMaaSSubscriptions.cy.ts
@@ -1,7 +1,7 @@
 import {
   ModelLocationSelectOption,
   ModelTypeLabel,
-} from '@odh-dashboard/model-serving/types/form-data';
+} from '@odh-dashboard/plugin-core/model-serving-labels';
 import {
   checkMaaSSubscriptionState,
   cleanupApiKeys,
