+++ packages/cypress/package.json
@@ -31,8 +31,8 @@
     "@odh-dashboard/maas": "*",
     "@odh-dashboard/nim-serving": "*",
     "@odh-dashboard/model-registry": "*",
-    "@odh-dashboard/model-serving": "*",
     "@odh-dashboard/model-training": "*",
+    "@odh-dashboard/plugin-core": "*",
     "@openshift/dynamic-plugin-sdk-utils": "^5.0.1",
     "@testing-library/cypress": "^10.0.1",
     "@testing-library/dom": "^9.3.4",
