+++ packages/model-serving/src/components/deploymentWizard/types.ts
@@ -53,32 +53,13 @@ export const isModelLocationType = (value?: string): value is ModelLocationType 
   return values.includes(value);
 };
 
-export enum ModelLocationSelectOption {
-  EXISTING = 'Existing connection',
-  PVC = 'Cluster storage',
-  S3 = 'S3 object storage',
-  OCI = 'OCI compliant registry',
-  URI = 'URI',
-}
-
-export enum ModelTypeLabel {
-  PREDICTIVE = 'Predictive model',
-  GENERATIVE = 'Generative AI model (Example, LLM)',
-}
-
-export enum ModelStateLabel {
-  STOPPED = 'Stopped',
-  STOPPING = 'Stopping',
-  STARTING = 'Starting',
-  READY = 'Ready',
-  RUNNING = 'Running',
-  FAILED_TO_LOAD = 'Failed to load',
-}
-
-export enum ModelStateToggleLabel {
-  START = 'Start',
-  STOP = 'Stop',
-}
+export {
+  ModelLocationSelectOption,
+  ModelTypeLabel,
+  ModelStateLabel,
+  ModelStateToggleLabel,
+  YAMLViewerToggleOption,
+} from '@odh-dashboard/plugin-core/model-serving-labels';
 
 export enum WizardStepTitle {
   PRECONFIGURE = 'Preconfigure deployment',
@@ -86,11 +67,6 @@ export enum WizardStepTitle {
   MODEL_DEPLOYMENT = 'Model deployment',
   ADVANCED_SETTINGS = 'Advanced settings',
   REVIEW = 'Review',
-}
-
-export enum YAMLViewerToggleOption {
-  YAML = 'YAML',
-  FORM = 'Form',
 }
 
 export type ModelLocationData = {
