+++ packages/plugin-core/package.json
@@ -14,7 +14,8 @@
     ".": "./src/core/index.ts",
     "./helpers/ui": "./src/core/helpers/ui.ts",
     "./extension-points": "./src/extension-points/index.ts",
-    "./testing": "./src/testing/utils.ts"
+    "./testing": "./src/testing/utils.ts",
+    "./model-serving-labels": "./src/model-serving-labels/index.ts"
   },
   "dependencies": {
     "@odh-dashboard/k8s-core": "*",
